обнова бэка

+++ portal/src/main/resources/db/migration/V12__create_taskComments_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS task_comments(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     comment VARCHAR(255) NOT NULL,
-    task_id SERIAL NOT NULL,
+    task_id BIGINT NOT NULL,
     client_id UUID,
     apteka_id INT
 );
